Cleaning of the test
