+++ test/unit/specs/fields/fieldRangeSlider.spec.js
@@ -1,5 +1,4 @@
 import { mount, createLocalVue } from "@vue/test-utils";
-import { checkAttribute2 } from "../util";
 
 import FieldRangeSlider from "src/fields/optional/fieldRangeSlider.vue";
 
@@ -65,7 +64,7 @@ describe("fieldRangeSlider.vue", () => {
 
 			attributes.forEach(name => {
 				it("should set " + name, () => {
-					checkAttribute2(name, wrapper, schema);
+					checkAttribute(name, wrapper, schema);
 				});
 			});
 		});
