+++ test/unit/specs/fields/fieldSpectrum.spec.js
@@ -1,5 +1,4 @@
 import { mount, createLocalVue } from "@vue/test-utils";
-import { checkAttribute2 } from "../util";
 
 import FieldSpectrum from "src/fields/optional/fieldSpectrum.vue";
 
@@ -53,7 +52,7 @@ describe("fieldSpectrum.vue", () => {
 
 			attributes.forEach(name => {
 				it("should set " + name, () => {
-					checkAttribute2(name, wrapper, schema);
+					checkAttribute(name, wrapper, schema);
 				});
 			});
 		});
