+++ test/unit/specs/fields/fieldSubmit.spec.js
@@ -1,5 +1,4 @@
 import { mount, createLocalVue } from "@vue/test-utils";
-import { checkAttribute2 } from "../util";
 
 import FieldSubmit from "src/fields/core/fieldSubmit.vue";
 
@@ -89,7 +88,7 @@ describe("fieldSubmit.vue", () => {
 
 			attributes.forEach(name => {
 				it("should set " + name, () => {
-					checkAttribute2(name, wrapper, schema);
+					checkAttribute(name, wrapper, schema);
 				});
 			});
 		});
