+++ test/unit/specs/fields/fieldSwitch.spec.js
@@ -1,5 +1,4 @@
 import { mount, createLocalVue } from "@vue/test-utils";
-import { checkAttribute2 } from "../util";
 
 import FieldSwitch from "src/fields/optional/fieldSwitch.vue";
 
@@ -51,7 +50,7 @@ describe("FieldSwitch.vue", () => {
 
 			attributes.forEach(name => {
 				it("should set " + name, () => {
-					checkAttribute2(name, wrapper, schema);
+					checkAttribute(name, wrapper, schema);
 				});
 			});
 		});
